Rename `banned_external_code` to `avoid_using_api`
