+++ example/analysis_options.yaml
@@ -19,7 +19,7 @@ custom_lint:
     - avoid_unnecessary_setstate
     - double_literal_format
     - avoid_unnecessary_type_assertions
-    - banned_external_code:
+    - avoid_using_api:
       severity: info
       entries:
         - class_name: Future
