+++ lib/solid_lints.dart
@@ -10,7 +10,7 @@ import 'package:solid_lints/lints/avoid_unnecessary_type_assertions/avoid_unnece
 import 'package:solid_lints/lints/avoid_unnecessary_type_casts/avoid_unnecessary_type_casts_rule.dart';
 import 'package:solid_lints/lints/avoid_unrelated_type_assertions/avoid_unrelated_type_assertions_rule.dart';
 import 'package:solid_lints/lints/avoid_unused_parameters/avoid_unused_parameters_rule.dart';
-import 'package:solid_lints/lints/banned_external_code/banned_external_code_rule.dart';
+import 'package:solid_lints/lints/avoid_using_api/avoid_using_api_rule.dart';
 import 'package:solid_lints/lints/cyclomatic_complexity/cyclomatic_complexity_metric.dart';
 import 'package:solid_lints/lints/double_literal_format/double_literal_format_rule.dart';
 import 'package:solid_lints/lints/function_lines_of_code/function_lines_of_code_metric.dart';
@@ -48,7 +48,7 @@ class _SolidLints extends PluginBase {
       AvoidUnnecessaryTypeCastsRule.createRule(configs),
       AvoidUnrelatedTypeAssertionsRule.createRule(configs),
       AvoidUnusedParametersRule.createRule(configs),
-      BannedExternalCodeRule.createRule(configs),
+      AvoidUsingApiRule.createRule(configs),
       NewlineBeforeReturnRule.createRule(configs),
       NoEmptyBlockRule.createRule(configs),
       NoEqualThenElseRule.createRule(configs),
