+++ lint_test/avoid_using_api/analysis_options.yaml
@@ -1,0 +1,7 @@
+analyzer:
+  plugins:
+    - ../custom_lint
+
+custom_lint:
+  rules:
+    - avoid_using_api:
