+++ lint_test/avoid_using_api/id_source_ban/lib/id_source_ban_test.dart
@@ -1,0 +1,15 @@
+void testingBannedCodeLint() {
+  final bannedCodeUsage = BannedCodeUsage();
+  bannedCodeUsage.test();
+
+  // expect_lint: avoid_using_api
+  test();
+}
+
+void test() {}
+
+class BannedCodeUsage {
+  BannedCodeUsage();
+
+  void test() {}
+}
