+++ lint_test/avoid_using_api/source_excludes_ban/lib/source_excludes_ban_test.dart
@@ -1,0 +1,6 @@
+import 'src/banned_library.dart';
+
+void testingBannedCodeLint() {
+  // expect_lint: avoid_using_api
+  banned;
+}
