+++ lint_test/avoid_using_api/source_excludes_ban/lib/src/domain/includes_test.dart
@@ -1,0 +1,9 @@
+import 'dart:collection';
+
+void testingBannedCodeLint() {
+  // expect_lint: avoid_using_api
+  final unmodifiable = UnmodifiableListView([1, 2, 3]);
+
+  // expect_lint: avoid_using_api
+  unmodifiable.first;
+}
