+++ lint_test/avoid_using_api/source_includes_ban/analysis_options.yaml
@@ -1,0 +1,18 @@
+analyzer:
+  plugins:
+    - ../../custom_lint
+
+custom_lint:
+  rules:
+    - avoid_using_api:
+      severity: warning
+      entries:
+        - source: dart:collection
+          reason: "dart:collection should not be used"
+          # Uses Glob notation, test using https://www.digitalocean.com/community/tools/glob
+          includes:
+            - "**/domain/**.dart"
+        - source: package:source_includes_ban/src/banned_library.dart
+          reason: "src/banned_library.dart should not be used anymore"
+          includes:
+            - "**/domain/**.dart"
