+++ lint_test/avoid_using_api/source_includes_ban/lib/src/domain/excludes_test.ex.dart
@@ -1,0 +1,8 @@
+import 'dart:collection';
+
+void testingBannedCodeLint() {
+  // expect_lint: avoid_using_api
+  final unmodifiable = UnmodifiableListView([1, 2, 3]);
+  // expect_lint: avoid_using_api
+  unmodifiable.first;
+}
