--- lint_test/banned_external_code/analysis_options.yaml
@@ -1,7 +1,0 @@
-analyzer:
-  plugins:
-    - ../custom_lint
-
-custom_lint:
-  rules:
-    - banned_external_code:
