--- lint_test/banned_external_code/id_source_ban/analysis_options.yaml
@@ -1,12 +1,0 @@
-analyzer:
-  plugins:
-    - ../../custom_lint
-
-custom_lint:
-  rules:
-    - banned_external_code:
-      severity: warning
-      entries:
-        - id: test
-          source: package:id_source_ban
-          reason: "test methods from package:id_source_ban is not allowed"
