--- lint_test/banned_external_code/id_source_ban/lib/id_source_ban_test.dart
@@ -1,15 +1,0 @@
-void testingBannedCodeLint() {
-  final bannedCodeUsage = BannedCodeUsage();
-  bannedCodeUsage.test();
-
-  // expect_lint: banned_external_code
-  test();
-}
-
-void test() {}
-
-class BannedCodeUsage {
-  BannedCodeUsage();
-
-  void test() {}
-}
