--- lint_test/banned_external_code/pubspec.yaml
@@ -1,16 +1,0 @@
-name: banned_external_code
-description: A starting point for Dart libraries or applications.
-version: 1.0.0
-publish_to: none
-
-environment:
-  sdk: ^3.1.3
-
-dependencies:
-  custom_lint: ^0.5.4
-
-dev_dependencies:
-  lints: ^3.0.0
-  test: ^1.21.0
-  solid_lints:
-    path: ../../
