--- lint_test/banned_external_code/source_ban/lib/source_ban_test.dart
@@ -1,6 +1,0 @@
-import 'src/banned_library.dart';
-
-void testingBannedCodeLint() {
-  // expect_lint: banned_external_code
-  banned;
-}
