--- lint_test/banned_external_code/source_excludes_ban/analysis_options.yaml
@@ -1,18 +1,0 @@
-analyzer:
-  plugins:
-    - ../../custom_lint
-
-custom_lint:
-  rules:
-    - banned_external_code:
-      severity: warning
-      entries:
-        - source: dart:collection
-          reason: "dart:collection should not be used"
-          # Uses Glob notation, test using https://www.digitalocean.com/community/tools/glob
-          excludes:
-            - "**.ex.dart"
-        - source: package:source_excludes_ban/src/banned_library.dart
-          reason: "src/banned_library.dart should not be used anymore"
-          excludes:
-            - "**.ex.dart"
