--- lint_test/banned_external_code/source_excludes_ban/lib/src/domain/includes_test.dart
@@ -1,9 +1,0 @@
-import 'dart:collection';
-
-void testingBannedCodeLint() {
-  // expect_lint: banned_external_code
-  final unmodifiable = UnmodifiableListView([1, 2, 3]);
-
-  // expect_lint: banned_external_code
-  unmodifiable.first;
-}
